Move most of the plugin to internal subdir.
The signer with address submodule is left at the root of the package so
it is easily accessible as
"@nomiclabs/hardhat-ethers/signer-with-address".
