+++ packages/hardhat-ethers/src/internal/type-extensions.ts
