+++ packages/hardhat-ethers/src/types/index.ts
@@ -1,6 +1,6 @@
 import type * as ethers from "ethers";
 
-import type { SignerWithAddress } from "../plugin/signer-with-address";
+import type { SignerWithAddress } from "../signer-with-address";
 
 export interface Libraries {
   [libraryName: string]: string;
